Add support for swiping in Espresso.
Introduces a swipeLeft() and swipeRight() action which simulates a swipe across
the view by injecting a DOWN event followed be a sequence of MOVE events, and
finally an UP event.

The swipe is implemented by a Swiper, with support for a 'fast' swipe (~100ms)
for normal use, and a 'slow' swipe (~1.5s) to aid when visually debugging a
test case.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=57073600